cp21: Refactor Unit test
Refactor Unit test

+++ Transenvios.Shipping.Api/Domains/ClientService/Client.cs
@@ -9,7 +9,7 @@ namespace Transenvios.Shipping.Api.Domains.ClientService
         public string? FirstName { get; set; }
         public string? LastName { get; set; }
         public string? Email { get; set; }
-        public long? CountryCode { get; set; }
+        public int? CountryCode { get; set; }
         public string? Phone { get; set; }
         public string? Role { get; set; }
         public string? PasswordHash { get; set; }
